took out unnecessary length declaration for Sequelize types

+++ database/models/trips.js
@@ -3,12 +3,12 @@ const db = require('../db');
 
 const Trips = db.define('trips', {
   id: {
-    type: Sequelize.INTEGER(11),
+    type: Sequelize.INTEGER,
     autoIncrement: true,
     primaryKey: true,
   },
   status: {
-    type: Sequelize.INTEGER(11),
+    type: Sequelize.INTEGER,
     allowNull: true,
   },
   route: {
@@ -16,7 +16,7 @@ const Trips = db.define('trips', {
     allowNull: true,
   },
   milesTraveled: {
-    type: Sequelize.INTEGER(11),
+    type: Sequelize.INTEGER,
     allowNull: true,
   },
 }, {
